requirements.txt
opencv-python>=4.6.0
pytorch>=1.12.0
torchvision>=0.13.0